+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - coverage checks ignored for branches named 'documentation' 
 - switch from Serveo to localtunnel
 - increased timeout from 5 to 8 minutes
+- generalized some variables such as repo owner, ci context
+- updated documentation
 
 ## [2.1.0]
 ### Added
@@ -21,7 +23,11 @@
 - suppress warnings about shadowing builtins in runAllTests
 - run tests in subfolders
 - filter out performance tests
+<<<<<<< HEAD
 - skip tests for commits to branches named 'documentation'
+=======
+- skip checks for commits to branches named 'documentation'
+>>>>>>> master
 
 ## [2.0.0]
 ### Added